src/day3/requirements.txt
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
chainlit>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
llama-index-core>=0.12.0
llama-index-vector-stores-qdrant>=0.4.0
llama-index-embeddings-fastembed>=0.3.0
fastembed>=0.4.0
llama-index-llms-anthropic>=0.6.0
qdrant-client>=1.12.0
